Add ActiveRecord specs

+++ spec/active_record/version_5_0_stable_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 # rubocop:disable all
-RSpec.describe 'ActiveRecord::Version51Stable' do
-  next unless Gem::Version.create('5.1') <= ActiveRecord.gem_version && ActiveRecord.gem_version < Gem::Version.create('5.2')
+RSpec.describe 'ActiveRecord::Version50Stable' do
+  next unless Gem::Version.create('5.0') <= ActiveRecord.gem_version && ActiveRecord.gem_version < Gem::Version.create('5.1')
 
   # 198bc1f785a7b826dfd50ccb068fdcbe463b34f6
   # Copy from rails/rails - activerecord/test/cases/dirty_test.rb
@@ -310,14 +310,6 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
       assert_equal ['arr', 'arr matey!'], pirate.catchphrase_change
     end
 
-    example 'test_virtual_attribute_will_change' do
-      assert_deprecated do
-        parrot = Parrot.create!(name: 'Ruby')
-        parrot.send(:attribute_will_change!, :cancel_save_from_callback)
-        assert parrot.has_changes_to_save?
-      end
-    end
-
     example 'test_association_assignment_changes_foreign_key' do
       pirate = Pirate.create!(catchphrase: 'jarl')
       pirate.parrot = Parrot.create!(name: 'Lorre')
@@ -358,13 +350,12 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
 
     example 'test_partial_update_with_optimistic_locking' do
       person = Person.new(first_name: 'foo')
+      old_lock_version = 1
 
       with_partial_writes Person, false do
         assert_queries(2) { 2.times { person.save! } }
         Person.where(id: person.id).update_all(first_name: 'baz')
       end
-
-      old_lock_version = person.lock_version
 
       with_partial_writes Person, true do
         assert_queries(0) { 2.times { person.save! } }
@@ -513,8 +504,8 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
       assert_equal 4, pirate.previous_changes.size
       assert_equal [nil, 'arrr'], pirate.previous_changes['catchphrase']
       assert_equal [nil, pirate.id], pirate.previous_changes['id']
-      assert_includes pirate.previous_changes, 'updated_on'
-      assert_includes pirate.previous_changes, 'created_on'
+      assert pirate.previous_changes.include?('updated_on')
+      assert pirate.previous_changes.include?('created_on')
       assert !pirate.previous_changes.key?('parrot_id')
 
       pirate.catchphrase = 'Yar!!'
@@ -576,24 +567,26 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
       travel_back
     end
 
-    class Testings < ActiveRecord::Base; end
-    example 'test_field_named_field' do
-      ActiveRecord::Base.connection.create_table :testings do |t|
-        t.string :field
-      end
-      assert_nothing_raised do
-        Testings.new.attributes
-      end
-    ensure
-      begin
+    if ActiveRecord::Base.connection.supports_migrations?
+      class Testings < ActiveRecord::Base; end
+      def test_field_named_field
+        ActiveRecord::Base.connection.create_table :testings do |t|
+          t.string :field
+        end
+        assert_nothing_raised do
+          Testings.new.attributes
+        end
+      ensure
+        begin
         ActiveRecord::Base.connection.drop_table :testings
       rescue StandardError
         nil
       end
-      ActiveRecord::Base.clear_cache!
+      end
     end
 
     example 'test_datetime_attribute_can_be_updated_with_fractional_seconds' do
+      skip 'Fractional seconds are not supported' unless subsecond_precision_supported?
       in_time_zone 'Paris' do
         target = Class.new(ActiveRecord::Base)
         target.table_name = 'topics'
@@ -622,7 +615,7 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
           jon = Person.create! first_name: 'Jon'
         end
 
-        assert ActiveRecord::SQLCounter.log_all.none? { |sql| sql.include?('followers_count') }
+        assert ActiveRecord::SQLCounter.log_all.none? { |sql| sql =~ /followers_count/ }
 
         jon.reload
         assert_equal 'Jon', jon.first_name
@@ -651,7 +644,7 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
       assert_equal('arrrr', pirate.catchphrase_was)
       assert pirate.catchphrase_changed?(from: 'arrrr')
       assert_not pirate.catchphrase_changed?(from: 'anything else')
-      assert_includes pirate.changed_attributes, :catchphrase
+      assert pirate.changed_attributes.include?(:catchphrase)
 
       pirate.save!
       pirate.reload
@@ -744,89 +737,6 @@ RSpec.describe 'ActiveRecord::Version51Stable' do
 
       person.first_name = nil
       assert person.changed?
-    end
-
-    example 'saved_change_to_attribute? returns whether a change occurred in the last save' do
-      person = Person.create!(first_name: 'Sean')
-
-      assert person.saved_change_to_first_name?
-      refute person.saved_change_to_gender?
-      assert person.saved_change_to_first_name?(from: nil, to: 'Sean')
-      assert person.saved_change_to_first_name?(from: nil)
-      assert person.saved_change_to_first_name?(to: 'Sean')
-      refute person.saved_change_to_first_name?(from: 'Jim', to: 'Sean')
-      refute person.saved_change_to_first_name?(from: 'Jim')
-      refute person.saved_change_to_first_name?(to: 'Jim')
-    end
-
-    example 'saved_change_to_attribute returns the change that occurred in the last save' do
-      person = Person.create!(first_name: 'Sean', gender: 'M')
-
-      assert_equal [nil, 'Sean'], person.saved_change_to_first_name
-      assert_equal [nil, 'M'], person.saved_change_to_gender
-
-      person.update(first_name: 'Jim')
-
-      assert_equal %w[Sean Jim], person.saved_change_to_first_name
-      assert_nil person.saved_change_to_gender
-    end
-
-    example 'attribute_before_last_save returns the original value before saving' do
-      person = Person.create!(first_name: 'Sean', gender: 'M')
-
-      assert_nil person.first_name_before_last_save
-      assert_nil person.gender_before_last_save
-
-      person.first_name = 'Jim'
-
-      assert_nil person.first_name_before_last_save
-      assert_nil person.gender_before_last_save
-
-      person.save
-
-      assert_equal 'Sean', person.first_name_before_last_save
-      assert_equal 'M', person.gender_before_last_save
-    end
-
-    example 'saved_changes? returns whether the last call to save changed anything' do
-      person = Person.create!(first_name: 'Sean')
-
-      assert person.saved_changes?
-
-      person.save
-
-      refute person.saved_changes?
-    end
-
-    example 'saved_changes returns a hash of all the changes that occurred' do
-      person = Person.create!(first_name: 'Sean', gender: 'M')
-
-      assert_equal [nil, 'Sean'], person.saved_changes[:first_name]
-      assert_equal [nil, 'M'], person.saved_changes[:gender]
-      assert_equal %w[id first_name gender created_at updated_at].sort, person.saved_changes.keys.sort
-
-      travel(1.second) do
-        person.update(first_name: 'Jim')
-      end
-
-      assert_equal %w[Sean Jim], person.saved_changes[:first_name]
-      assert_equal %w[first_name lock_version updated_at].sort, person.saved_changes.keys.sort
-    end
-
-    example 'changed? in after callbacks returns true but is deprecated' do
-      klass = Class.new(ActiveRecord::Base) do
-        self.table_name = 'people'
-
-        after_save do
-          ActiveSupport::Deprecation.silence do
-            raise 'changed? should be true' unless changed?
-          end
-          raise 'has_changes_to_save? should be false' if has_changes_to_save?
-        end
-      end
-
-      person = klass.create!(first_name: 'Sean')
-      refute person.changed?
     end
 
     private
